+++ lib/deimos/backends/kafka_async.rb
@@ -4,9 +4,10 @@ module Deimos
   module Backends
     # Backend which produces to Kafka via an async producer.
     class KafkaAsync < Base
+
       # :nodoc:
       def self.execute(producer_class:, messages:)
-        Karafka.producer.produce_many_async(messages)
+        Deimos.producer_for(producer_class.topic).produce_many_async(messages)
       end
     end
   end