+++ lib/deimos/utils/outbox_producer.rb
@@ -207,7 +207,7 @@ module Deimos
 
         batch[current_index..-1].in_groups_of(batch_size, false).each do |group|
           @logger.debug("Publishing #{group.size} messages to #{@current_topic}")
-          Karafka.producer.produce_many_sync(group)
+          Deimos.producer_for(@current_topic).produce_many_sync(group)
           current_index += group.size
           @logger.info("Sent #{group.size} messages to #{@current_topic}")
         end